requirements.txt
torch
torchvision
timm
numpy
pandas
matplotlib
pillow
scikit-learn
tqdm
